+++ com.unity.ml-agents/CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to
 - Fixed a null reference exception that occurred when loading an ONNX model file that was generated with a new
 version of the Python trainer (0.26.0 or newer). (#5350)
 - Added checked to prevent training with incompatible versions of the python trainer (0.26.0 or newer). (#5370)
+### Minor Changes
+#### ml-agents / ml-agents-envs / gym-unity (Python)
+- Added minimal analytics collection to LL-API (#5511)
 
 ## [1.0.7] - 2021-03-04
 ### Minor Changes